İd 11 hane şartı tamamlandı.

+++ src/admin.java
@@ -1,15 +1,15 @@
 public class admin extends adminPanel{
-    int id;
+    long id;
     String name;
     int password;
 
-    public admin(int id, String name, int password) {
+    public admin(long id, String name, int password) {
         this.id = id;
         this.name = name;
         this.password = password;
     }
 
-    boolean enterAdmin(int id,int password){
+    boolean enterAdmin(long id,int password){
         if (this.id == id && this.password == password) {
             return true;
         }
