+++ src/patient.java
@@ -44,7 +44,7 @@ public class patient {
 
     }
 
-    boolean enterPatient(int id){
+    boolean enterPatient(long id){
         if (this.id == id) {
             return true;
         }
